🙀 rename the project
