+++ .pre-commit-config.yaml
@@ -1,20 +1,30 @@
-files: jaxtro
+files: gwkokab
 repos:
-  - repo: https://github.com/psf/black
-    rev: 24.1.0
-    hooks:
-      - id: black
-        args: [ --line-length=120 ]
-  - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: "v0.2.0"
-    hooks:
-      - id: ruff
-        args: [ "--fix" ]
-      - id: ruff-format # formatter
-        types_or: [ python, pyi, jupyter ]
-  - repo: https://github.com/RobertCraigie/pyright-python
-    rev: v1.1.349
-    hooks:
-      - id: pyright
-        additional_dependencies:
-          [ jax, jaxlib, jaxtyping, tfp-nightly, tqdm, jaxampler, configargparse, mplcursors, rift ]
+    - repo: https://github.com/psf/black
+      rev: 24.1.0
+      hooks:
+          - id: black
+            args: [--line-length=120]
+    - repo: https://github.com/astral-sh/ruff-pre-commit
+      rev: "v0.2.0"
+      hooks:
+          - id: ruff
+            args: ["--fix"]
+          - id: ruff-format # formatter
+            types_or: [python, pyi, jupyter]
+    - repo: https://github.com/RobertCraigie/pyright-python
+      rev: v1.1.349
+      hooks:
+          - id: pyright
+            additional_dependencies:
+                [
+                    jax,
+                    jaxlib,
+                    jaxtyping,
+                    tfp-nightly,
+                    tqdm,
+                    jaxampler,
+                    configargparse,
+                    mplcursors,
+                    rift,
+                ]
