+++ CITATION.bib
@@ -1,7 +1,7 @@
-@software{jaxtro2023github,
+@software{gwkokab2023github,
     author  = {Meesum Qazalbash, Muhammad Zeeshan, Richard O'Shaughnessy},
-    title   = {{jaxtro}: A JAX-based gravitational-wave population inference},
-    url     = {https://github.com/gwkokab/jaxtro},
+    title   = {{gwkokab}: A JAX-based gravitational-wave population inference},
+    url     = {https://github.com/gwkokab/gwkokab},
     version = {0.0.3},
     year    = {2023}
 }
