+++ gwkokab/_src/utils/__init__.py
@@ -1,10 +1,10 @@
-#  Copyright 2023 The Jaxtro Authors
+#  Copyright 2023 The GWKokab Authors
 #
 #  Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
 #  You may obtain a copy of the License at
 #
-#     http://www.apache.org/licenses/LICENSE-2.0
+#      http://www.apache.org/licenses/LICENSE-2.0
 #
 #  Unless required by applicable law or agreed to in writing, software
 #  distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,15 +12,14 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-
 from __future__ import annotations
 
-from jaxampler.rvs import TruncNormal
-from jaxtyping import Array
-
-from .abstractmodel import AbstractModel
-
-
-class AbstractEccentricityModel(AbstractModel):
-    def add_error(self, x: Array, scale: float = 0.5, size: int = 10) -> Array:
-        return TruncNormal(loc=x, scale=scale, low=0.0, high=0.5).rvs(shape=(size,))
+from .misc import (
+    chirp_mass as chirp_mass,
+    dump_configurations as dump_configurations,
+    mass_ratio as mass_ratio,
+    reduced_mass as reduced_mass,
+    symmetric_mass_ratio as symmetric_mass_ratio,
+)
+from .parser import cmd_parser as cmd_parser, parse_config as parse_config
+from .popgen import PopulationGenerator as PopulationGenerator
