+++ gwkokab/utils.py
@@ -1,4 +1,4 @@
-#  Copyright 2023 The Jaxtro Authors
+#  Copyright 2023 The GWKokab Authors
 #
 #  Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
@@ -14,6 +14,13 @@
 
 from __future__ import annotations
 
-from .utils import interpolate as interpolate, interpolate_hdf5 as interpolate_hdf5, load_hdf5 as load_hdf5
-from .vt_from_mass import vt_from_mass as vt_from_mass
-from .vt_from_mass_spin import vt_from_mass_spin as vt_from_mass_spin
+from gwkokab._src.utils import (
+    chirp_mass as chirp_mass,
+    cmd_parser as cmd_parser,
+    dump_configurations as dump_configurations,
+    mass_ratio as mass_ratio,
+    parse_config as parse_config,
+    PopulationGenerator as PopulationGenerator,
+    reduced_mass as reduced_mass,
+    symmetric_mass_ratio as symmetric_mass_ratio,
+)
