+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools]
-py-modules = ["jaxtro"]
+py-modules = ["gwkokab"]
 
 [tool.setuptools.packages]
 find = { }
@@ -14,8 +14,8 @@ authors = [
     { name = "Muhammad Zeeshan", email = "m.zeeshan5885@gmail.com" },
     { name = "Richard O'Shaughnessy", email = "richardoshaughnessyrossma@gmail.com" },
 ]
-name = "jaxtro"
-version = "0.0.3.dev"
+name = "gwkokab"
+version = "0.0.1.dev"
 maintainers = [
     { name = "Meesum Qazalbash", email = "meesumqazalbash@gmail.com" },
     { name = "Muhammad Zeeshan", email = "m.zeeshan5885@gmail.com" },
@@ -64,16 +64,16 @@ keywords = [
 ]
 
 [project.urls]
-Homepage = "https://github.com/gwkokab/jaxtro"
-Issues = "https://github.com/gwkokab/jaxtro/issues"
-Discussion = "https://github.com/gwkokab/jaxtro/discussions"
-Changelog = "https://github.com/gwkokab/jaxtro/blob/master/CHANGELOG.md"
+Homepage = "https://github.com/gwkokab/gwkokab"
+Issues = "https://github.com/gwkokab/gwkokab/issues"
+Discussion = "https://github.com/gwkokab/gwkokab/discussions"
+Changelog = "https://github.com/gwkokab/gwkokab/blob/master/CHANGELOG.md"
 
 [project.scripts]
-jaxtro_genie = "jaxtro.main:main"
-vt_mass_spin = "jaxtro.vt_cli:vt_mass_spin"
-vt_mass = "jaxtro.vt_cli:vt_mass"
-vt_mass_grid = "jaxtro.vt_grid:vt_mass_grid"
+gwkokab_genie = "gwkokab.main:main"
+vt_mass_spin = "gwkokab.vt_cli:vt_mass_spin"
+vt_mass = "gwkokab.vt_cli:vt_mass"
+vt_mass_grid = "gwkokab.vt_grid:vt_mass_grid"
 
 
 [tool.ruff]
@@ -91,4 +91,4 @@ order-by-type = false
 
 [tool.pyright]
 reportIncompatibleMethodOverride = true
-include = ["jaxtro", "tests"]
+include = ["gwkokab", "tests"]
